+++ core/src/main/java/cn/ttplatform/wh/cmd/factory/GetResultCommandSerializer.java
@@ -29,7 +29,7 @@ public class GetResultCommandSerializer extends AbstractDistributableSerializer 
     }
 
     @Override
-    public int getFactoryType() {
+    public byte getFactoryType() {
         return DistributableType.GET_COMMAND_RESULT;
     }
