+++ core/src/main/java/cn/ttplatform/wh/cmd/factory/SetResultCommandSerializer.java
@@ -29,7 +29,7 @@ public class SetResultCommandSerializer extends AbstractDistributableSerializer 
     }
 
     @Override
-    public int getFactoryType() {
+    public byte getFactoryType() {
         return DistributableType.SET_COMMAND_RESULT;
     }
