+++ server/src/main/java/cn/ttplatform/wh/message/RequestVoteMessage.java
@@ -1,11 +1,7 @@
 package cn.ttplatform.wh.message;
 
 import cn.ttplatform.wh.constant.DistributableType;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
+import lombok.*;
 import lombok.experimental.SuperBuilder;
 
 /**
@@ -25,7 +21,7 @@ public class RequestVoteMessage extends AbstractMessage{
     private int lastLogTerm;
 
     @Override
-    public int getType() {
+    public byte getType() {
         return DistributableType.REQUEST_VOTE;
     }
